move applyAppStateMiddleware to store folder
