+++ app/lib/store/index.ts
@@ -3,7 +3,7 @@ import createSagaMiddleware from 'redux-saga';
 
 import reducers from '../../reducers';
 import sagas from '../../sagas';
-import applyAppStateMiddleware from '../appStateMiddleware';
+import applyAppStateMiddleware from './appStateMiddleware';
 
 let sagaMiddleware;
 let enhancers;
